refactor: organize tests under test folder
