--- jest-integration.config.ts
@@ -1,14 +1,0 @@
-import type {Config} from 'jest';
-
-const config: Config = {
-  setupFilesAfterEnv: ['jest-extended/all'],
-  testEnvironment: 'node',
-  roots: ['<rootDir>/integration'],
-  testMatch: ['**/*.test.ts'],
-  transform: {
-    '^.+\\.tsx?$': 'ts-jest',
-  },
-  testTimeout: 30000,
-};
-
-export default config;
