+++ jest.config.ts
@@ -8,6 +8,7 @@ const config: Config = {
   transform: {
     '^.+\\.tsx?$': 'ts-jest',
   },
+  testTimeout: 30000,
 };
 
 export default config;
