+++ test/integration/get-set-delete.test.ts
@@ -7,7 +7,7 @@ import {
   CacheSetIfNotExists,
   MomentoErrorCode,
   SimpleCacheClient,
-} from '../src';
+} from '../../src';
 import {TextEncoder} from 'util';
 import {
   SetupIntegrationTest,
@@ -16,7 +16,7 @@ import {
   ItBehavesLikeItValidatesCacheName,
   WithCache,
 } from './integration-setup';
-import {sleep} from '../src/internal/utils/sleep';
+import {sleep} from '../../src/internal/utils/sleep';
 
 const {Momento, IntegrationTestCacheName} = SetupIntegrationTest();
