+++ test/unit/auth/credential-provider.test.ts
@@ -1,4 +1,4 @@
-import {CredentialProvider} from '../../src/auth/credential-provider';
+import {CredentialProvider} from '../../../src/auth/credential-provider';
 
 const testToken =
   'eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJmb29Abm90LmEuZG9tYWluIiwiY3AiOiJjb250cm9sLXBsYW5lLWVuZHBvaW50Lm5vdC5hLmRvbWFpbiIsImMiOiJjYWNoZS1lbmRwb2ludC5ub3QuYS5kb21haW4ifQo.rtxfu4miBHQ1uptWJ2x3UiAwwJYcMeYIkkpXxUno_wIavg4h6YJStcbxk32NDBbmJkJS7mUw6MsvJNWaxfdPOw';
