+++ test/unit/utils/jwt.test.ts
@@ -1,5 +1,5 @@
-import {decodeJwt} from '../../src/internal/utils/jwt';
-import {InvalidArgumentError} from '../../src';
+import {decodeJwt} from '../../../src/internal/utils/jwt';
+import {InvalidArgumentError} from '../../../src';
 
 const TEST_AUTH_TOKEN_NO_ENDPOINT =
   'eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJpbnRlZ3JhdGlvbiJ9.ZOgkTs';
